requirements.txt
requests==2.32.4
PyYAML==6.0.2
transformers==4.55.0
numpy==2.3.1
tqdm==4.67.1